Refactored StatusPagerTabStripViewController
